+++ src/main/java/be/baur/sda/Node.java
@@ -1,30 +1,47 @@
 package be.baur.sda;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.function.Predicate;
 
 /**
  * A {@code Node} is an object in a tree-like structure that stems from a single
  * root node. A node may have child nodes (in which case it is called a parent
- * node), or it may be a leaf node. Nodes usually have information attached, the
- * nature of which depends on the implementation.
- * 
- * @see AbstractNode
+ * node). Nodes usually have information attached, the nature of which depends
+ * on the implementation.
  */
-public interface Node {
-
+public interface Node<T extends Node<T>> {
 
 	/**
-	 * Returns the name of this node. Typically, a node name is an identifier by
-	 * which it can be found or addressed. A node name does not have to be unique.
-	 * <p>
-	 * This method should not return null. The default method returns the simple
-	 * name of the class implementing this interface.
+	 * Returns an <i>unmodifiable</i> list of child nodes, which may be empty. This
+	 * method never returns null.
+	 * 
+	 * @return a list of nodes, not null
+	 */
+	List<T> nodes();
+
+	/**
+	 * Returns the parent of this node or null if it has no parent.
+	 * 
+	 * @implSpec assumes proper F-bounded polymorphism usage; subclasses should bind
+	 *           T to their own type, or this method may not return an object of
+	 *           type T (in which case the caller should verify type safety).
+	 * 
+	 * @return the parent node, may be null
+	 */
+	T getParent();
+
+	// Default methods below this line
+
+	/**
+	 * Returns the name by which this node can be found or addressed. A node name
+	 * does not have to be unique and can be empty, but should not be null.
+	 * 
+	 * @implNote the default method returns the simple name of the class
+	 *           implementing this interface, e.g. the result of
+	 *           {@link Class#getSimpleName}.
 	 * 
 	 * @return the node name, not null
-	 * @see Class#getSimpleName
 	 */
 	default String getName() {
 		return getClass().getSimpleName();
@@ -32,46 +49,23 @@ public interface Node {
 
 
 	/**
-	 * Returns the parent of this node or null if it has no parent.
-	 * 
-	 * @param <T> the type of node
-	 * @return the parent node, may be null
-	 */
-	<T extends Node> T getParent();
-
-
-	/**
 	 * Returns the ultimate ancestor of this node. This method returns the node
 	 * itself if it has no parent (in which case it <i>is</i> the root node).
 	 * 
-	 * @param <T> the type of node
-	 * @return the root node, not null, may be this node
-	 */
-	@SuppressWarnings("unchecked")
-	default <T extends Node> T root() {
-		final Node parent = getParent();
-		return ((parent != null) ? parent.root() : (T) this);
-	}
-
-
-	/**
-	 * Returns an <i>unmodifiable</i> list of child nodes, which may be empty. This
-	 * method never returns null.
+	 * @implNote relies on proper F-bounded polymorphism usage; subclasses should
+	 *           bind T to their own type, or this method may not return an object
+	 *           of type T (in which case the caller should verify type safety).
 	 * 
-	 * @param <T> the type of node
-	 * @return a list, not null
+	 * @return the root node, not null, may be this
 	 */
-	<T extends Node> List<T> nodes();
-
-
-	/**
-	 * Returns true if this node has no child nodes.
-	 * 
-	 * @return true or false
-	 * @see #isParent
-	 */
-	default boolean isLeaf() {
-		return nodes().isEmpty();
+	default T root() {
+	    Node<?> current = this;
+		for (Node<?> p; (p = current.getParent()) != null; ) {
+	        current = p;
+	    }
+	    @SuppressWarnings("unchecked")
+	    T root = (T) current;
+	    return root;
 	}
 
 
@@ -79,7 +73,6 @@ public interface Node {
 	 * Returns true if this node has one or more child nodes.
 	 * 
 	 * @return true or false
-	 * @see #isLeaf
 	 */
 	default boolean isParent() {
 		return ! nodes().isEmpty();
@@ -87,40 +80,15 @@ public interface Node {
 
 
 	/**
-	 * Adds a child node to this node. This method takes after the add method of the
-	 * {@code Collection} interface and implementations should strive to comply with
-	 * the specified requirements.
-	 * 
-	 * @see Collection#add(Object)
-	 * 
-	 * @param node the node to be added
-	 * @return true if this node changed as a result of the call
-	 */
-	boolean add(Node node);
-
-
-	/**
-	 * Removes a child node from this node. This method takes after the remove
-	 * method of the {@code Collection} interface and implementations should strive
-	 * to comply with the specified requirements.
-	 * 
-	 * @see Collection#remove(Object)
-	 * 
-	 * @param node the node to be removed
-	 * @return true if this node changed as a result of the call
-	 */
-	boolean remove(Node node);
-
-
-	/**
 	 * Returns the first child node with the specified name, or null if no such node
-	 * is found. This method uses the result of {@link #getName} to find a match.
+	 * is found.
 	 * 
-	 * @param <T> the type of node
+	 * @implNote this method uses the result of {@link #getName} to find a match.
+	 * 
 	 * @param name a node name
 	 * @return a node, may be null
 	 */
-	default <T extends Node> T get(String name) {
+	default T get(String name) {
 		return get(n -> n.getName().equals(name));
 	}
 
@@ -129,29 +97,27 @@ public interface Node {
 	 * Returns the first child node that satisfy the given predicate, or null if no
 	 * such node is found.
 	 * 
-	 * @param <T> the type of node
 	 * @param predicate a boolean valued function of one argument
 	 * @return a node, may be null
 	 */
-	@SuppressWarnings("unchecked")
-	default <T extends Node> T get(Predicate<? super Node> predicate) {
-		for (Node node : nodes())
+	default T get(Predicate<? super T> predicate) {
+		for (T node : nodes())
 			if (predicate.test(node))
-				return (T) node;
+				return node;
 		return null;
 	}
 
 
 	/**
 	 * Returns a list of child nodes with the specified name, or an empty list if no
-	 * such nodes are found. This method uses the result of {@link #getName} to find
-	 * matching nodes.
+	 * such nodes are found.
 	 * 
-	 * @param <T> the type of node
+	 * @implNote this method uses the result of {@link #getName} to find a match.
+	 * 
 	 * @param name a node name
 	 * @return a list, not null
 	 */
-	default <T extends Node> List<T> getAll(String name) {
+	default List<T> getAll(String name) {
 		return getAll(n -> n.getName().equals(name));
 	}
 
@@ -160,16 +126,14 @@ public interface Node {
 	 * Returns a list of child nodes that satisfy the given predicate, or an empty
 	 * list if no such nodes are found.
 	 * 
-	 * @param <T> the type of node
 	 * @param predicate a boolean valued function of one argument
 	 * @return a list, not null
 	 */
-	@SuppressWarnings("unchecked")
-	default <T extends Node> List<T> getAll(Predicate<? super Node> predicate) {
-		List<T> list = new ArrayList<T>();
-		for (Node node : nodes())
+	default List<T> getAll(Predicate<? super T> predicate) {
+		var list = new ArrayList<T>();
+		for (T node : nodes())
 			if (predicate.test(node))
-				list.add((T) node);
+				list.add(node);
 		return list;
 	}
 
@@ -179,39 +143,60 @@ public interface Node {
 	 * empty list if no such nodes are found. In the resulting list, matching child
 	 * nodes are returned before matching sibling nodes (and their children).
 	 * 
-	 * @param <T> the type of node
 	 * @param predicate a boolean valued function of one argument
 	 * @return a list, not null
 	 */
-	@SuppressWarnings("unchecked")
-	default <T extends Node> List<T> find(Predicate<? super Node> predicate) {
-		List<T> list = new ArrayList<T>();
-		for (Node node : nodes()) {
+	default List<T> find(Predicate<? super T> predicate) {
+		var list = new ArrayList<T>();
+		for (T node : nodes()) {
 			if (predicate.test(node))
-				list.add((T) node);
+				list.add(node);
 			list.addAll(node.find(predicate));
 		}
 		return list;
 	}
 
+
 	/**
 	 * Returns the location of this node in X-path style. If a node occurs more than
 	 * once in the same context, its position is indicated in square brackets.
-	 * <p>
-	 * The path is constructed from node names returned by {@link #getName}, and may
-	 * contain whatever characters are allowed in these names.
+	 * 
+	 * @implNote the resulting path is constructed from node names returned by
+	 *           {@link #getName}, and may contain any character allowed in these
+	 *           names. If name is empty, an asterisk (*) is used instead.
 	 * 
 	 * @return the path to this node, for example {@code /root/message[2]/text[1]}
 	 */
 	default String path() {
 		
-		final String name = getName();
-		final Node parent = getParent();
-		final List<Node> same = (parent != null) ? parent.getAll(name) : null;
-		final int pos = (same != null && same.size() > 1) ? same.indexOf(this)+1 : 0;
+		T parent = getParent(); 
+		
+		String name = getName();
+		if (name.isEmpty()) name = "*"; // not necessarily a DataNode
+		
+		if (parent == null) // start of the path, so return
+			return "/" + name;
+				
+		var sameNodes = parent.getAll(name);
+		// position in list of nodes with the same name
+		final int pos = sameNodes.size() > 1 ? position(sameNodes) : 0;
 
-		return (parent != null ? parent.path() : "")
-			+ "/" + name + (pos > 0 ? "[" + pos + "]" : "");
+		return parent.path() + "/" + name + (pos > 0 ? "[" + pos + "]" : "");
+	}
+
+
+	/*
+	 * Private helper that returns the position of this node in the provided list,
+	 * in the range {@code 1 .. list.size()} or 0 if the list does not contain it.
+	 * The reason we use this helper rather then list.indexOf() is because we want
+	 * an identity compare rather than equals(), which might be overridden.
+	 * Precondition: the list must not be null.
+	 */
+	private int position(List<T> list) {
+		for (int i = 0, n = list.size(); i < n; ++i)
+			if (list.get(i) == this)
+				return ++i;
+		return 0;
 	}
 
 }