+++ src/main/java/be/baur/sda/io/Parser.java
@@ -18,7 +18,7 @@ import be.baur.sda.Node;
  * 
  * @see SDAParser
  */
-public interface Parser<T extends Node> {
+public interface Parser<T extends Node<?>> {
 
 	/**
 	 * Creates a node from a character input stream.
@@ -32,14 +32,14 @@ public interface Parser<T extends Node> {
 
 
 	/**
-	 * Creates a node from a file with SDA content.
+	 * Creates a node from a file with SDA content, assuming UTF-8 encoding.
 	 * 
 	 * @param file an input file, not null
 	 * @return a (root) node, never null
-	 * @throws IOException    if an I/O operation failed
-	 * @throws ParseException if a parsing error occurs
+	 * @throws IOException        if an I/O operation failed
+	 * @throws FileParseException if a parsing error occurs
 	 */
-	default T parse(File file) throws IOException, ParseException {
+	default T parse(File file) throws IOException, FileParseException {
 
 		Objects.requireNonNull(file, "input file must not be null");
 		try (
@@ -53,7 +53,7 @@ public interface Parser<T extends Node> {
 			throw new IOException("error reading from " + file, e);
 		}
 		catch (ParseException e) {
-			throw new FileParseException(file.toString(), e);
+			throw new FileParseException(file.getPath(), e);
 		}
 	}