+++ src/test/java/test/Test.java
@@ -1,25 +1,27 @@
 package test;
 
 
+import java.io.File;
 import java.util.function.Function;
 
 /** A convenience class with testing methods that accept Lambda expressions */
 public class Test {
 
-	Function<String, String> strfun;
-	String prefix;
+	private Function<String, String> function;
+	private String prefix;
+	private int failures;
 	
-	public Test(Function<String, String> strfun, String prefix) {
-		this.strfun = strfun; this.prefix = prefix;
+	public Test(Function<String, String> function, String prefix) {
+		this.function = function; this.prefix = prefix;
 	}
 	
-	public Test(Function<String, String> strfun) {
-		this(strfun, "");
+	public Test(Function<String, String> function) {
+		this(function, "");
 	}
 	
-	public void ts1(String scenario, String str, String expected) {
+	public void s(String scenario, String str, String expected) {
 		
-		String result = strfun.apply(str);
+		String result = function.apply(str);
 
 		if (expected == null) expected = str;
 		expected = prefix + expected;
@@ -30,26 +32,19 @@ public class Test {
 			System.out.println("\n" + scenario + " FAILED!");
 			System.out.println("    EXPECTED: " + expected);
 			System.out.println("    RETURNED: " + result);
+			++failures;
 		}
 	}
 
+	public void checkFailures() throws Exception {
+		
+		if (failures > 0)
+			throw new Exception("Failed tests: " + failures);
+	}
 	
-//	public void testError(String scenario, String input, String expected) {
-//		
-//		try { 
-//			strfun.apply(input);
-//		}
-//		catch (Exception e) { 
-//			if (e.getMessage().equals(expected)) 
-//				System.out.print(scenario + " ");
-//			else {
-//				System.out.println("\n" + scenario + " FAILED!");
-//				System.out.println("    EXPECTED: " + expected);
-//				System.out.println("    RETURNED: " + e.getMessage());
-//			}
-//			return;
-//		}
-//		System.out.println(scenario + " FAILED - exception expected");
-//	}
+	// convenience method to load a resource file
+	public static File getResourceFile(String name) {
+		return new File(Test.class.getResource(name).getFile());
+	}
 
 }